move development code to spearate file
